Reorganise js file structure
Fix cookie banner local port
Fix escaping of single quotes in messages
