--- app/uk/gov/hmrc/trackingconsentfrontend/assets/src/bundle.ts
@@ -1,9 +1,0 @@
-import '../styles/styles.scss'
-import renderBanner from './ui/renderBanner'
-import userPreferenceFactory from './userPreferenceFactory'
-import enableGtm from './enableGtm'
-
-enableGtm()
-document.addEventListener('DOMContentLoaded', () => {
-  renderBanner(userPreferenceFactory())
-})
