+++ app/uk/gov/hmrc/trackingconsentfrontend/assets/src/common/castToArray.ts
