+++ app/uk/gov/hmrc/trackingconsentfrontend/assets/src/entrypoints/servicePage.ts
@@ -1,0 +1,9 @@
+import '../../styles/styles.scss'
+import renderBanner from '../ui/renderBanner'
+import userPreferenceFactory from '../domain/userPreferenceFactory'
+import enableGtm from '../interfaces/gtm'
+
+enableGtm()
+document.addEventListener('DOMContentLoaded', () => {
+  renderBanner(userPreferenceFactory())
+})
