+++ app/uk/gov/hmrc/trackingconsentfrontend/assets/src/entrypoints/settingsPage.ts
@@ -1,0 +1,4 @@
+import settingsFormHandler from '../ui/settingsFormHandler'
+import userPreferenceFactory from '../domain/userPreferenceFactory'
+
+settingsFormHandler(document, userPreferenceFactory())
