--- app/uk/gov/hmrc/trackingconsentfrontend/assets/src/settingsPage.ts
@@ -1,4 +1,0 @@
-import settingsFormHandler from './settingsFormHandler'
-import userPreferenceFactory from './userPreferenceFactory'
-
-settingsFormHandler(document, userPreferenceFactory())
