+++ conf/application.conf
@@ -109,5 +109,5 @@ assets {
 }
 
 cookie-banner {
-    port = 9000
+    port = 12345
 }
