+++ conf/messages
@@ -3,14 +3,13 @@ testPage.title = Service test page
 
 cookieSettings.intro.paragraph1 = Cookies are files saved on your phone, tablet or computer when you visit a website
 cookieSettings.intro.paragraph2 = We use cookies to store information about how you use HMRC services, such as the pages you visit.
-cookieSettings.intro.paragraph3 = We use 4 types of cookie. You can choose which cookies you're happy for us to use.\
-        
+cookieSettings.intro.paragraph3 = We use 4 types of cookie. You can choose which cookies you''re happy for us to use.
 cookieSettings.usage.title = Cookies that measure website use
 cookieSettings.usage.paragraph1 = We use Google Analytics to measure how you use the website so we can improve it based on user needs. We do not allow Google to use or share the data about how you use this site.
 cookieSettings.usage.paragraph2 = Google Analytics sets cookies that store anonymised information about:
 cookieSettings.usage.bullet1 = how you got to the site
 cookieSettings.usage.bullet2 = the pages you visit on HMRC services and how long you spend on each page
-cookieSettings.usage.bullet3 = what you click on while you're visiting the site
+cookieSettings.usage.bullet3 = what you click on while you''re visiting the site
 cookieSettings.usage.settinglabel.on = Use cookies that measure my website use
 cookieSettings.usage.settinglabel.off = Do not use cookies that measure my website use
 cookieSettings.usage.settingid = usage
